+++ src/importexport/tabledit/internal/readinglist.cpp
@@ -122,7 +122,7 @@ void ReadingList::analyze()
             m_status[m_list.at(itemsUsed - 1).last].repeatEnd = true;
             if (itemsUsed > 2) {
                 // check interaction with alternative endings
-                m_status[m_list.at(itemsUsed - 1).last].ending = itemsUsed;
+                m_status[m_list.at(itemsUsed - 1).last].ending = static_cast<int>(itemsUsed);
             }
             m_list.erase(m_list.begin(), m_list.begin() + itemsUsed);
             continue;
